+++ CHANGELOG.md
@@ -1,2 +1,5 @@
+# Unreleased (0.1.1)
+* Add `Batch::pull_waiting_items`.
+
 # 0.1.0
 * Implement `BatchMutex`.